AUTH_TEST/cursor/cursor3/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Authentication and security
passlib==1.7.4
bcrypt==4.0.1
PyJWT==2.8.0
python-multipart==0.0.6

# Testing
pytest==7.4.3
httpx==0.25.1

